Moved contents in RecyclerViewMergeAdapter_ into RecyclerViewMergeAdapter and removed RecyclerViewSubAdapter as it was a redundant class. Eliminated generics as much as possible as any RecyclerView.Adapter should be able to part of the merge adapter's data set.
